Renamed base package from net.korti.bettermuffling to io.korti.bettermuffling. Optimized imports.
