--- src/main/java/net/korti/bettermuffling/client/gui/GuiHandler.java
@@ -1,21 +1,0 @@
-package net.korti.bettermuffling.client.gui;
-
-import net.korti.bettermuffling.client.ClientProxy;
-import net.korti.bettermuffling.common.tileentity.TileMuffling;
-import net.minecraft.client.Minecraft;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.math.BlockPos;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
-
-@OnlyIn(Dist.CLIENT)
-public final class GuiHandler {
-
-    public static void openMufflingGui(BlockPos pos) {
-        final TileEntity te = ClientProxy.getWorld().getTileEntity(pos);
-        if(te instanceof TileMuffling) {
-            Minecraft.getInstance().displayGuiScreen(new MufflingBlockGui((TileMuffling) te));
-        }
-    }
-
-}
